Update choosing-antenna.rst
line 27:
from ...will double range
to: what was put

added line 28 to 30:
people are using transmitters illegally and have been causing interference to stations with proper authority to transmit. Also some folks, including my area, are using video and telemetry transmitters in the Amateur bands, and are not licensed to do so. In all local cases, their equipment was causing interference to licensed stations and unwilling to cease operations.

I'd be pretty sad to learn of transmissions for legal stations causing someone's UAS to crash as a result of the UAS operator not knowing or refusing stop using equipment causing interference. 

Not to sound arrogant, I've learned a few things working with radios and antennas for the last 21 years.